# api/kitchen.openapi.yaml
openapi: 3.1.0
info:
  title: Авито.Кухня API
  version: 0.1.0
  description: |
    Public API будущего web-клиента и закрытый Partner API заведений.

    `X-User-ID` используется только как MVP-заглушка идентификации и не является
    механизмом безопасности. Денежные значения передаются целым числом копеек,
    единственная валюта MVP — RUB.
servers:
  - url: http://localhost:8080
    description: Локальный Docker Compose
...
paths:
...
  /api/v1/restaurants:
    get:
      tags: [Public catalog]
      operationId: listRestaurants
      summary: Получить список заведений
      description: Возвращает заведения в стабильном порядке с cursor-pagination.
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Страница заведений.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantPage'
              examples:
                default:
                  $ref: '#/components/examples/RestaurantList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/restaurants/{restaurantId}:
    get:
      tags: [Public catalog]
      operationId: getRestaurant
      summary: Получить карточку заведения
      security: []
      parameters:
        - $ref: '#/components/parameters/RestaurantID'
      responses:
        '200':
          description: Карточка заведения.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
              examples:
                default:
                  $ref: '#/components/examples/RestaurantDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/restaurants/{restaurantId}/menu:
    get:
      tags: [Public catalog]
      operationId: getRestaurantMenu
      summary: Получить опубликованное меню заведения
      description: Доступность является копией платформы; окончательный остаток проверяет ресторан.
      security: []
      parameters:
        - $ref: '#/components/parameters/RestaurantID'
      responses:
        '200':
          description: Текущая опубликованная версия меню.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
              examples:
                default:
                  $ref: '#/components/examples/PublishedMenu'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders:
    post:
      tags: [Public orders]
      operationId: createOrder
      summary: Создать заказ
      description: |
        Сервер проверяет принадлежность позиций одному заведению, актуальные цены и
        копию доступности, затем сам рассчитывает итог. Заказ, снимки позиций,
        начальная история и outbox-событие создаются одной транзакцией.

        Первый успешный запрос возвращает `201`. Повтор с тем же `Idempotency-Key`
        и эквивалентным телом возвращает существующий заказ с `200`. Тот же ключ
        с другим телом возвращает `409 IDEMPOTENCY_KEY_REUSED`.
      security: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              default:
                $ref: '#/components/examples/CreateOrderRequest'
      responses:
        '201':
          description: Заказ создан.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
            Location:
              description: URI созданного заказа.
              schema:
                type: string
                example: /api/v1/orders/aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
            Idempotency-Replayed:
              description: Всегда `false` для впервые созданного заказа.
              schema:
                type: boolean
                example: false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                created:
                  $ref: '#/components/examples/PendingOrder'
        '200':
          description: Идемпотентный повтор; возвращён ранее созданный заказ.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
            Idempotency-Replayed:
              description: Всегда `true` для идемпотентного повтора.
              schema:
                type: boolean
                example: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                replayed:
                  $ref: '#/components/examples/PendingOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Запрос конфликтует с текущими данными или прежним idempotency key.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                priceChanged:
                  $ref: '#/components/examples/PriceChangedError'
                itemUnavailable:
                  $ref: '#/components/examples/ItemUnavailableError'
                restaurantClosed:
                  $ref: '#/components/examples/RestaurantClosedError'
                idempotencyKeyReused:
                  $ref: '#/components/examples/IdempotencyKeyReusedError'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          description: Синтаксически корректный заказ нарушает правило одного заведения.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                multipleRestaurants:
                  $ref: '#/components/examples/MultipleRestaurantsError'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Public orders]
      operationId: listUserOrders
      summary: Получить заказы пользователя
      description: Возвращает только заказы `X-User-ID`, от новых к старым.
      security: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Страница заказов пользователя.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
              examples:
                default:
                  $ref: '#/components/examples/UserOrderList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}:
    get:
      tags: [Public orders]
      operationId: getUserOrder
      summary: Получить заказ пользователя
      description: Чужой заказ маскируется ответом `404 NOT_FOUND`.
      security: []
      parameters:
        - $ref: '#/components/parameters/OrderID'
        - $ref: '#/components/parameters/UserID'
      responses:
        '200':
          description: Заказ со снимками позиций и историей статусов.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                default:
                  $ref: '#/components/examples/PendingOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/cancel:
    post:
      tags: [Public orders]
      operationId: cancelUserOrder
      summary: Отменить заказ пользователя
      description: |
        Отмена возможна только из `pending_confirmation`, пока worker не начал
        передачу ресторану. Статус, история и остановка ожидающего outbox-события
        фиксируются атомарно.
      security: []
      parameters:
        - $ref: '#/components/parameters/OrderID'
        - $ref: '#/components/parameters/UserID'
      responses:
        '200':
          description: Заказ отменён.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                cancelled:
                  $ref: '#/components/examples/CancelledOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Заказ уже передаётся или его статус не допускает отмену.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notCancellable:
                  $ref: '#/components/examples/OrderNotCancellableError'
        '500':
          $ref: '#/components/responses/InternalError'
  /partner/v1/menu:
    put:
      tags: [Partner menu]
      operationId: publishPartnerMenu
      summary: Опубликовать полную версию меню
      description: |
        Заведение определяется Bearer API-ключом. Версия в теле должна быть больше
        текущей. Точный повтор уже применённой версии с тем же содержимым возвращает
        прежний результат; старая версия или то же число с другим содержимым дают
        `409 STALE_MENU_VERSION`.
      security:
        - PartnerBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishMenuRequest'
            examples:
              default:
                $ref: '#/components/examples/PublishMenuRequest'
      responses:
        '200':
          description: Меню опубликовано или возвращён результат точного повтора.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuPublication'
              examples:
                default:
                  $ref: '#/components/examples/MenuPublication'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Версия меню устарела или повторно использована с другим содержимым.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                staleVersion:
                  $ref: '#/components/examples/StaleMenuVersionError'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
  /partner/v1/menu/items/{externalItemId}:
    patch:
      tags: [Partner menu]
      operationId: patchPartnerMenuItem
      summary: Изменить опубликованную позицию меню
      description: |
        `expected_version` должна совпасть с текущей версией меню. После успешного
        изменения платформа атомарно увеличивает версию на один.
      security:
        - PartnerBearer: []
      parameters:
        - $ref: '#/components/parameters/ExternalItemID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchMenuItemRequest'
            examples:
              availability:
                $ref: '#/components/examples/PatchMenuItemRequest'
      responses:
        '200':
          description: Позиция изменена, версия меню увеличена.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemMutation'
              examples:
                default:
                  $ref: '#/components/examples/MenuItemMutation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: "`expected_version` не совпадает с текущей версией."
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                staleVersion:
                  $ref: '#/components/examples/StaleMenuVersionError'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
  /partner/v1/orders:
    get:
      tags: [Partner orders]
      operationId: listPartnerOrders
      summary: Получить заказы заведения
      description: Заведение определяется Bearer API-ключом; сортировка — от новых к старым.
      security:
        - PartnerBearer: []
      parameters:
        - $ref: '#/components/parameters/OrderStatusFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Страница заказов заведения.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerOrderPage'
              examples:
                default:
                  $ref: '#/components/examples/PartnerOrderList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /partner/v1/orders/{orderId}/decision:
    post:
      tags: [Partner orders]
      operationId: decidePartnerOrder
      summary: Принять или отклонить ожидающий заказ
      description: Допустим только переход из `pending_confirmation` в `accepted` или `rejected`.
      security:
        - PartnerBearer: []
      parameters:
        - $ref: '#/components/parameters/OrderID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDecisionRequest'
            examples:
              accept:
                $ref: '#/components/examples/OrderDecisionRequest'
      responses:
        '200':
          description: Решение применено.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerOrder'
              examples:
                accepted:
                  $ref: '#/components/examples/PartnerAcceptedOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Текущий статус не допускает указанное решение.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidTransition:
                  $ref: '#/components/examples/InvalidStatusTransitionError'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    PartnerBearer:
      type: http
      scheme: bearer
      bearerFormat: Partner API key
      description: Статический высокоэнтропийный ключ; в БД платформы хранится только его хеш.
  headers:
    XRequestID:
      description: Идентификатор запроса, также присутствующий в теле ошибки.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        example: req-01J9XYZ7E7T1R6M4B8N3P2Q5S6
  parameters:
    RestaurantID:
      name: restaurantId
      in: path
      required: true
      description: Внутренний UUID заведения на платформе.
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      description: Максимальное число элементов страницы.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      description: Непрозрачный cursor из предыдущего ответа; нельзя конструировать на клиенте.
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
  responses:
    BadRequest:
      description: Запрос не прошёл синтаксическую или структурную валидацию.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            message: Поле items должно содержать хотя бы одну позицию.
            details:
              field: items
            request_id: req-01J9XYZ7E7T1R6M4B8N3P2Q5S6
...
    InternalError:
      description: Неожиданная внутренняя ошибка без раскрытия деталей реализации.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: INTERNAL_ERROR
            message: Внутренняя ошибка сервиса.
            request_id: req-01J9XYZ7E7T1R6M4B8N3P2Q5S6
...
  schemas:
...
    Restaurant:
      type: object
      additionalProperties: false
      required: [id, name, is_accepting_orders, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        address:
          type: string
          maxLength: 500
        is_accepting_orders:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RestaurantPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/RestaurantSummary'
        next_cursor:
          type: string
          minLength: 1
          maxLength: 512
...
  examples:
    RestaurantList:
      value:
        items:
          - id: 11111111-1111-4111-8111-111111111111
            name: Demo Bistro
            description: Обеды и выпечка
            is_accepting_orders: true
        next_cursor: eyJuYW1lIjoiRGVtbyBCaXN0cm8ifQ
...